Removed HOCS for focused container
